requirements.txt
azure-identity==1.14.0
Flask==2.3.2
openai==0.27.7
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0
langchain==0.0.244
streamlit==1.25.0
PyPDF2==3.0.1
tiktoken==0.4.0
PyMuPDF==1.22.5
unstructured[pdf]